move master tcp stuff into crate::tcp
