+++ dnp3/src/master/serial.rs
@@ -1,7 +1,7 @@
 use crate::app::ExponentialBackOff;
 use crate::master::handle::{Listener, MasterConfig, MasterHandle};
 use crate::master::session::{MasterSession, RunError};
-use crate::master::ClientState;
+use crate::tcp::ClientState;
 use crate::transport::TransportReader;
 use crate::transport::TransportWriter;
 use crate::util::task::Shutdown;
